Transportando object orcamentoCotacaoEmailQueue para mensageria, gravar...

See merge request ar-clube/ar-clube!50

+++ Global/Infra/InfraBanco/Modelos/TorcamentoCotacaoMensagem.cs
@@ -53,7 +53,7 @@ namespace InfraBanco.Modelos
         public DateTime DataHoraCadastro { get; set; }
 
         [Column("IdOrcamentoCotacaoEmailQueue")]
-        public int? IdOrcamentoCotacaoEmailQueue { get; set; }
+        public long? IdOrcamentoCotacaoEmailQueue { get; set; }
         
         [Column("PendenciaTratada")]
         public bool? PendenciaTratada { get; set; }
